backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
httpx>=0.24.1
google-generativeai>=0.8.2
python-dotenv>=1.0.1
groq>=0.11.0
